test/requirements.txt
scikit-image==0.17.2
silence-tensorflow==1.1.1
pydot-ng==2.0.0